requirements.txt
langchain>=0.1.13
langchain-community>=0.0.13
langchain-google-genai>=0.0.8
google-generativeai>=0.3.2
duckduckgo-search>=3.9
python-dotenv>=1.0.1
streamlit>=1.35.0
